requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.0
pydantic-settings==2.7.0
httpx==0.28.0
python-jose[cryptography]==3.3.0
redis==5.2.0
websockets==14.1
python-multipart==0.0.17
pyyaml==6.0.2
